:recycle: refactor: Add Category Domain
