+++ src/main/kotlin/com/team1/mvp_test/domain/category/dto/CreateCategoryRequest.kt
@@ -1,4 +1,4 @@
-package com.team1.mvp_test.domain.mvptest.dto
+package com.team1.mvp_test.domain.category.dto
 
 data class CreateCategoryRequest(
     val category: String
