+++ src/main/kotlin/com/team1/mvp_test/domain/category/repository/CategoryRepository.kt
@@ -1,6 +1,6 @@
-package com.team1.mvp_test.domain.mvptest.repository
+package com.team1.mvp_test.domain.category.repository
 
-import com.team1.mvp_test.domain.mvptest.model.Category
+import com.team1.mvp_test.domain.category.model.Category
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface CategoryRepository : JpaRepository<Category, Long> {
